llmcoolproject/requirements.txt
Django
daphne
langchain
langchain-community
channels
python-dotenv
aiohttp
django-cors-headers
openai